Revert "Feature/fragment action handling"
