+++ activity/src/main/java/com/cube/fusion/android/activity/actions/DefaultActivityActionHandlers.kt
@@ -2,14 +2,17 @@ package com.cube.fusion.android.activity.actions
 
 import android.app.Activity
 import android.content.Intent
+import android.net.Uri
 import android.view.View
+import com.cube.fusion.android.activity.R
 import com.cube.fusion.android.activity.WebViewActivity
-import com.cube.fusion.android.core.actions.DefaultAndroidActionHandlers
 import com.cube.fusion.android.core.actions.FusionAndroidActionHandler
-import com.cube.fusion.android.core.actions.MutableListMultiActionHandler
-import com.cube.fusion.android.core.utils.extensions.handledByView
+import com.cube.fusion.android.core.actions.MultiActionHandler
+import com.cube.fusion.core.model.action.EmailAction
 import com.cube.fusion.core.model.action.LinkAction
+import com.cube.fusion.core.model.action.NativeAction
 import com.cube.fusion.core.model.action.PageAction
+import com.cube.fusion.core.utils.Constants
 
 /**
  * [FusionAndroidActionHandler] implementation for the default handling of actions by redirects to activities
@@ -20,13 +23,23 @@ import com.cube.fusion.core.model.action.PageAction
  *
  * @param pageIntentGenerator a method to generate intents for [PageAction]s to redirect to
  */
-class DefaultActivityActionHandlers(pageIntentGenerator: (View, PageAction) -> Intent?) : MutableListMultiActionHandler() {
+class DefaultActivityActionHandlers(pageIntentGenerator: (View, PageAction) -> Intent?) : MultiActionHandler() {
 	/**
 	 * @param pageClass the activity class for [PageAction]s to redirect to
 	 */
 	constructor(pageClass: Class<out Activity>) : this({ view, _ -> Intent(view.context, pageClass) })
 
 	companion object {
+		/**
+		 * Convenience method to handle a nullable intent created during action handling by starting an activity if not-null
+		 *
+		 * @param view the view to handle the nullable intent with
+		 * @return true if the intent is non-null, and therefore startActivity has been called; false otherwise
+		 */
+		fun Intent?.handledByView(view: View) = this?.let {
+			view.context.startActivity(it)
+			true
+		} ?: false
 
 		/**
 		 * Get a default handler for [PageAction]s that redirects based on the intent generated by [pageIntentGenerator] for each [PageAction], passing the page URL as an extra
@@ -41,12 +54,42 @@ class DefaultActivityActionHandlers(pageIntentGenerator: (View, PageAction) -> I
 		}
 
 		/**
-		 * Get a default handler for in-app [LinkAction]s; i.e ones with [LinkAction.inApp] true
-		 * @return a [FusionAndroidActionHandler] that redirects [LinkAction]s to [WebViewActivity] if inApp
+		 * Get a default handler for [LinkAction]s
+		 * @return a [FusionAndroidActionHandler] that redirects [LinkAction]s either to [WebViewActivity] if inApp, or an external app otherwise
 		 */
-		fun defaultInAppLinkActivityActionHandler() = FusionAndroidActionHandler { view, action ->
-			if (action is LinkAction && action.inApp) {
-				WebViewActivity.getIntent(view.context, action.extractClick()).handledByView(view)
+		fun defaultLinkActivityActionHandler() = FusionAndroidActionHandler { view, action ->
+			if (action is LinkAction) {
+				val intent = if (action.inApp) {
+					WebViewActivity.getIntent(view.context, action.extractClick())
+				}
+				else {
+					Intent(Intent.ACTION_VIEW).apply {
+						data = Uri.parse(action.extractClick())
+					}
+				}
+				intent.handledByView(view)
+			}
+			else false
+		}
+
+		/**
+		 * Get a default handler for [EmailAction]s
+		 * @return a [FusionAndroidActionHandler] that handles [EmailAction]s by launching an email app selector with the relevant email extras
+		 */
+		fun defaultEmailActivityActionHandler() = FusionAndroidActionHandler { view, action ->
+			if (action is EmailAction) {
+				val selectorIntent = Intent(Intent.ACTION_SENDTO).apply {
+					data = Uri.parse(Constants.MAIL_TO)
+				}
+				val emailIntent = Intent(Intent.ACTION_SEND).apply {
+					putExtra(Intent.EXTRA_EMAIL, action.to)
+					putExtra(Intent.EXTRA_BCC, action.bcc)
+					putExtra(Intent.EXTRA_CC, action.cc)
+					putExtra(Intent.EXTRA_SUBJECT, action.subject)
+					putExtra(Intent.EXTRA_TEXT, action.body)
+					selector = selectorIntent
+				}
+				Intent.createChooser(emailIntent, view.context.getString(R.string.send_email_header)).handledByView(view)
 			}
 			else false
 		}
@@ -54,8 +97,36 @@ class DefaultActivityActionHandlers(pageIntentGenerator: (View, PageAction) -> I
 
 	override val childHandlers = mutableListOf(
 		defaultPageActivityActionHandler(pageIntentGenerator),
-		defaultInAppLinkActivityActionHandler(),
-		DefaultAndroidActionHandlers.defaultExternalLinkActivityActionHandler(),
-		DefaultAndroidActionHandlers.defaultEmailActivityActionHandler()
+		defaultLinkActivityActionHandler(),
+		defaultEmailActivityActionHandler()
 	)
+
+	/**
+	 * Register a new handler for [NativeAction]s
+	 *
+	 * @param handler the handling method to call for any native action.
+	 * 	Should return true if the action is successfully handled, and false otherwise.
+	 */
+	fun registerNativeAction(handler: (View, NativeAction) -> Boolean) {
+		childHandlers.add(FusionAndroidActionHandler { view, action ->
+			if (action is NativeAction) {
+				handler(view, action)
+			}
+			else false
+		})
+	}
+
+	/**
+	 * Register a new handler for [NativeAction]s with a specific link, which returns true and calls [onLinkMatch] whenever the action link exactly matches [link]
+	 *
+	 * @param link the exact link to handle
+	 * @param onLinkMatch the method to call whenever a [NativeAction] with the given link is handled
+	 */
+	fun registerNativeActionForLink(link: String, onLinkMatch: (View, NativeAction) -> Unit) = registerNativeAction { view, nativeAction ->
+		if (nativeAction.link == link) {
+			onLinkMatch(view, nativeAction)
+			true
+		}
+		else false
+	}
 }
